Refactor Dexie data handling and update type definitions for improved clarity
- Replaced the deprecated DexieRepository with a new structure in dexieTypes and dexieUtils, enhancing type safety and organization.
- Updated hooks to utilize the new dexieTypes for better type management across the application.
- Refactored offer-related logic to streamline data fetching and state management, improving overall performance.
- Removed the obsolete useTickerData hook and integrated its functionality into useCatTokens, ensuring a cleaner codebase.
- Adjusted TypeScript imports and definitions for consistency and clarity throughout the project.
